Add files via upload
reciprocated_pairs.cpp uses an unordered_map to determine whether there are nodes in a directed graph that are configured in such a manner as to have reciprocal communication (i.e. node A points to node B AND node B points to node A). However, due to the nature of the underlying data structure and the use of its key-value pairs, it is currently written in such a way as to allow only unique pairs to exist. In other words, the current solution is applicable to directed graphs but not to multi-edge directed graphs. For instance, if node A has two edges pointing towards node B AND node B has two edges pointing to node A, such a configuration (arrangement of an object's elements) will not be counted as an extra pair.

       - A solution to this is to use a hash table or to simply construct the map using key-value pairs that allow for the increase of the value
         of an unordered_map's elements if there are multiple such elements in the provided graph.