Count total elements of Trie
Provides the same result as using the get member function Trie::size() but uses a stack to traverse the Trie based on Depth-First-Search algorithm. Need to develop a function that displays all elements of the Trie or an iterator that helps start at any element from the root and begin traversal or display from there.