Implement overloaded insertion operator function to display elements in Trie using object itself
Using a friend function that has access to the members of the Trie data structure allows for access to the root of the Trie data structure. This allows the invocation of a recursive member function that can be called to modify the ostream object cout to append to its output stream all elements of the trie data structure.

The functionality is demonstrated with the use of a simple cout statement that uses the << insertion operator and the Trie object instance alone.